requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
httpx>=0.24.0

# Kubernetes
kubernetes>=27.2.0

# LLM Providers
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# Optional: For quantization (install if you have compatible GPU)
# bitsandbytes>=0.39.0

# Optional: Azure OpenAI
# azure-openai>=1.0.0

# Web interface (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0

# CLI
click>=8.1.0
rich>=13.0.0
bitsandbytes>=0.39.0

# Development/testing (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0